Added ingest command
A `telepresence ingest` command, similar to `telepresence intercept`,
provides local access to the volume mounts and environment variables of
a targeted container. However, unlike `intercept`, `ingest` does not
redirect traffic to the container and ensures that the mounted volumes
are read-only.

An ingest requires a traffic-agent to be installed in the pods of the
targeted workload. Beyond that, it's a client-side operation. This
allows developers to have multiple simultaneous ingests on the same
container.